requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==6.0
django-cloudinary-storage==0.3.0
gunicorn==20.1.0
idna==3.11
packaging==25.0
pillow==12.0.0
python-decouple==3.8
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
whitenoise==6.11.0
